requirements.txt
Flask==0.10.1
Flask-Login==0.2.11
Flask-Mail==0.9.0
Flask-Principal==0.4.0
Flask-Restless==0.13.1
Flask-SQLAlchemy==1.0
Flask-Security==1.7.3
Flask-WTF==0.10.0
Jinja2==2.7.2
MarkupSafe==0.19
MySQL-python==1.2.5
SQLAlchemy==0.9.6
SQLAlchemy-Utils==0.26.9
WTForms==2.0.1
WTForms-Alchemy==0.12.8
WTForms-Components==0.9.5
Werkzeug==0.9.4
blinker==1.3
decorator==3.4.0
geopy==0.99
gnureadline==6.3.3
infinity==1.3
intervals==0.3.1
ipython==2.1.0
itsdangerous==0.23
mimerender==0.5.4
passlib==1.6.2
python-dateutil==2.2
python-mimeparse==0.1.4
simplejson==3.3.3
six==1.7.3
validators==0.6.0
wsgiref==0.1.2
